blog_scraper/requirements.txt
flask==2.3.3
sqlalchemy==2.0.23
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
feedparser==6.0.10
openai==1.3.7
spacy==3.7.2
newspaper3k==0.2.8
python-dotenv==1.0.0
sentence-transformers==2.2.2
chromadb==0.4.18
openpyxl==3.1.2
pandas==2.1.4
httpx==0.25.2
tensorflow==2.15.0
trafilatura==1.7.0
readability-lxml==0.8.1
requests-cache==1.2.0
fake-useragent==1.5.1
langdetect==1.0.9
